Extract common YAML logic out to its own module
Without originally realizing it, I put classes in
"kobweb-common" which expose one of its dependencies, "kaml"
(a Kotlin-serialization library for YAML), in some of its APIs
of some of its classes, despite not declaring it as an api
dependency.

I noticed this when another consumer of this library, the
Kobweb CLI, included its own copy of the "kaml" library.
However, when I tried to upgrade its version, the JVM exploded
at runtime because some code was refactored between the version
"common" is using and the version the CLI is using.

Really, the proper thing to do is declare an API declaration if
the API is forcing users to use a particular library. It didn't
feel right to put that inside "kobweb-common", though, which
contains a handful of useful utilities in their own right.

As a compromise, we moved the subset of YAML classes to a new module,
so third party programs can consider using "kobweb-common" without
YAML stuff, and if they do depend on "kobweb-common-yaml", they do
so with the understanding that the `kaml` dependency will be
hoisted upon their codebase.