Test: Export in parallel
The risk of this feature is probably not worth it, given that
in my experiment it does shave significant time off of the page
snapshotting (like, 30%), but that task is a fairly small part
of the whole overall expert (so, we save 3 seconds on 1 minute,
for example).

Could be worth revisiting if maybe this is a bigger benefit for
larger sites.

For me, I tend to export and walk away (or it's done on a CI).

Still, if this is worth pursuing, the following still needs to
be done:

* Configurability (on / off? how many max threads to export at
  the same time?)
* Better logging (right now, what's happening each each route
  immediately logs that it's starting, a bunch of time passes,
  and then you get a bunch of "export took ####ms" messages.)
  I didn't see it but you might even get printlns overlapping
  each other? Anyway, there should probably be a dedicated
  coroutine for consuming information and handling the printlns
  separately.